backend/pyproject.toml
[project]
name = "pi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "gdal==3.11.3",
    "numpy>=2.3.3",
    "pymongo[srv]>=4.15.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyvips[binary]>=3.0.0",
]
[tool.pyright]
venvPath="."
venv = ".venv"
